requirements.txt
paho-mqtt>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
flask>=3.0.0
pytest>=7.0.0
